pentra_core/requirements-local.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.29,<1
redis>=5.0,<6
httpx>=0.27,<1
openai>=1.107.0,<2
anthropic>=0.60.0,<1
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
PyYAML>=6.0,<7
alembic>=1.13,<2
docker>=7.0,<8
pytest>=8.0,<9
